+++ FadeBasic/Tests/TokenVm_GC.cs
@@ -91,6 +91,52 @@ ENDFUNCTION
         Assert.That(vm.heap.Allocations, Is.EqualTo(allocationCount));
     }
 
+    [Test]
+    public void GC_ArraysDontRuinCopiedOutData()
+    {
+        var src = @"
+type egg
+ z
+endtype
+
+global x as egg
+global x2 as egg
+for n = 0 to 2
+    dim a(5) as egg
+    a(3) = { z = 50+n }
+    if (n = 0)
+        x = a(3)
+    endif
+    if (n = 2)
+        x2 = a(3)
+    endif
+next
+
+q = x.z
+q2 = x2.z
+
+x3 = a(3)
+q3 = x3.z
+";
+        Setup(src, out var compiler, out var prog);
+        
+        var vm = new VirtualMachine(prog);
+        vm.hostMethods = compiler.methodTable;
+
+        vm.Execute2(0);
+        vm.heap.Sweep();
+
+        var v = vm.dataRegisters[6];
+        var v2 = vm.dataRegisters[7];
+        var v3 = vm.dataRegisters[9];
+        
+        Assert.That(v, Is.EqualTo(50));
+        Assert.That(v2, Is.EqualTo(52));
+        Assert.That(v3, Is.EqualTo(52));
+        
+        Assert.That(vm.heap.Allocations, Is.EqualTo(4)); // left over array, 3 eggs
+    }
+
     [TestCase(@"
 for n = 1 to 100
     ` concat of literals is fine